feat: get & put detail log #36

+++ backend/src/main/java/com/example/backend/search/repository/SearchRepository.java
@@ -4,6 +4,6 @@ import com.example.backend.search.domain.CaseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface SearchRepository extends JpaRepository<CaseEntity, Long>, JpaSpecificationExecutor<CaseEntity> {
+public interface SearchRepository extends JpaRepository<CaseEntity, Integer>, JpaSpecificationExecutor<CaseEntity> {
 }   // JpaSpecificationExecutor => Specification을 이용한 동적 검색
